Set unfinished deployments to 'failed' on boot
If there are deployments in the state 'active' or 'new' when booting up,
that probably means that Applikatoni crashed while these deployments
were running.

If they are still marked as 'new'/'active' after booting up that
prevents us from creating new deployments ("another deployment to this
target is already in progress").

This change adds a simple SQL statement and executes it after
establishing a database connection to set all deployments in the state
'new'/'active' to 'failed', since we can't be sure that they succeeded.