use markdown for flowdock notifier message
Flowdock recently added markdown support to their messages, so we can make the deployment notification a little bit prettier.

It still contains the same amount of information.

I did not use the "```" code formatting, because this will lead to ugly text highlighting, when entering normal text (which will be likely the case in the comment section).

see http://blog.flowdock.com/2015/08/12/markdown-support-and-an-emoji-picker-are-here/