Fix race-condition in summary generation
Before this change the summaries generated by the `generateSummary`
function would sometimes contain the correct status and sometimes not.

The reason was that the function uses `deployment.State` to determine
whether the deployment was successful or not. But the problem here is
that the deployment state may or may not be updated to the correct state
when the notifiers are run. That's because the LogEntry is sent before
`manager.Start()` returns here:

  https://github.com/applikatoni/applikatoni/blob/2f90c09e12452d3c29ba381313f0b6c422fa3adb/server/handlers.go#L274-L279

Sometimes the deployment state is updated in the database before the
notifiers are run and sometimes not. A race-condition.

With this fix we use the LogEntry itself to determine if the deployment
was successful or not. The rest of the information we get by loading
the deployment from the database is already persisted, so it's safe to
use it.