Use a database busy timeout of 30sec
The reason for this commit is this issue:

    https://github.com/applikatoni/applikatoni/issues/35

We ran into "database is locked" errors because of webhooks trying to
read from the database on every log entry they received -- all the while
are log entries being saved to the database.

This change increases the busy timeout of the 5sec default value to
30sec. We'll see how that plays out. First tests confirm that it might
slow the deployment process down but the errors disappear since the
queries are retried.

For more information, start reading here:

* https://github.com/applikatoni/applikatoni/issues/35
* https://github.com/mattn/go-sqlite3/issues/274
* https://gist.github.com/mrnugget/0eda3b2b53a70fa4a894