Fix "deployment already in progress" check
Before this change the `activeDeploymentExists` function would return
true if a deployment for another application was active with the same
target name. That's wrong. The check needs to be scoped to the
application level, so that multiple applications can be deployed at the
same time, but not one application to different targets at the same
time.